kayle_be/requirements.txt
amqp==5.0.2
asgiref==3.3.1
bidict==0.21.2
billiard==3.6.3.0
blinker==1.4
Brotli==1.0.9
celery @ https://github.com/celery/celery/tarball/master
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.2.1
demjson==2.2.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
gevent==20.9.0
gevent-websocket==0.10.1
gevent-websocket2==1.0
greenlet==0.4.17
gunicorn==20.0.4
h11==0.11.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kaitaistruct==0.9
kombu==5.0.2
ldap3==2.8.1
MarkupSafe==1.1.1
mitmproxy==5.3.0
msgpack==1.0.0
passlib==1.7.4
prompt-toolkit==3.0.8
protobuf==3.13.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.20
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.1
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2020.4
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.20
tornado==6.1
urllib3==1.25.11
urwid==2.1.2
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wsproto==0.15.0
zope.event==4.5.0
zope.interface==5.2.0
zstandard==0.14.0
